// klink-web/src/generated/models/QueryExistingPayload.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Klink API
 * API to manage Klinks
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { QueryExistingPayloadKlinksInner } from './QueryExistingPayloadKlinksInner';
import {
    QueryExistingPayloadKlinksInnerFromJSON,
    QueryExistingPayloadKlinksInnerFromJSONTyped,
    QueryExistingPayloadKlinksInnerToJSON,
} from './QueryExistingPayloadKlinksInner';

/**
 * 
 * @export
 * @interface QueryExistingPayload
 */
export interface QueryExistingPayload {
    /**
     * 
     * @type {Array<QueryExistingPayloadKlinksInner>}
     * @memberof QueryExistingPayload
     */
    klinks: Array<QueryExistingPayloadKlinksInner>;
}

/**
 * Check if a given object implements the QueryExistingPayload interface.
 */
export function instanceOfQueryExistingPayload(value: object): value is QueryExistingPayload {
    if (!('klinks' in value) || value['klinks'] === undefined) return false;
    return true;
}

export function QueryExistingPayloadFromJSON(json: any): QueryExistingPayload {
    return QueryExistingPayloadFromJSONTyped(json, false);
}

export function QueryExistingPayloadFromJSONTyped(json: any, ignoreDiscriminator: boolean): QueryExistingPayload {
    if (json == null) {
        return json;
    }
    return {
        
        'klinks': ((json['klinks'] as Array<any>).map(QueryExistingPayloadKlinksInnerFromJSON)),
    };
}

export function QueryExistingPayloadToJSON(value?: QueryExistingPayload | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'klinks': ((value['klinks'] as Array<any>).map(QueryExistingPayloadKlinksInnerToJSON)),
    };
}