// klink-web/src/generated/models/QueryExistingPayloadKlinksInner.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Klink API
 * API to manage Klinks
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface QueryExistingPayloadKlinksInner
 */
export interface QueryExistingPayloadKlinksInner {
    /**
     * 
     * @type {string}
     * @memberof QueryExistingPayloadKlinksInner
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof QueryExistingPayloadKlinksInner
     */
    readKey: string;
}

/**
 * Check if a given object implements the QueryExistingPayloadKlinksInner interface.
 */
export function instanceOfQueryExistingPayloadKlinksInner(value: object): value is QueryExistingPayloadKlinksInner {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('readKey' in value) || value['readKey'] === undefined) return false;
    return true;
}

export function QueryExistingPayloadKlinksInnerFromJSON(json: any): QueryExistingPayloadKlinksInner {
    return QueryExistingPayloadKlinksInnerFromJSONTyped(json, false);
}

export function QueryExistingPayloadKlinksInnerFromJSONTyped(json: any, ignoreDiscriminator: boolean): QueryExistingPayloadKlinksInner {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'readKey': json['readKey'],
    };
}

export function QueryExistingPayloadKlinksInnerToJSON(value?: QueryExistingPayloadKlinksInner | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'id': value['id'],
        'readKey': value['readKey'],
    };
}